+++ tests/unit/ska_tmc_cdm/serialisation/subarray_node/configure/test_receptorgroup.py
@@ -1,6 +1,7 @@
 from ska_tmc_cdm import CODEC
 from ska_tmc_cdm.messages.subarray_node.configure.receptorgroup import (
     FixedTrajectory,
+    TableTrajectory,
 )
 from tests.utils import assert_json_is_equal
 
@@ -29,4 +30,34 @@ class TestFixedTrajectory:
         )
         assert instance == CODEC.loads(
             FixedTrajectory, TestFixedTrajectory.FULL_JSON
+        )
+
+
+class TestTableTrajectory:
+    FULL_JSON = """
+    {
+        "name": "table",
+        "attrs": {
+            "x": [1.1, 2.2, 3.3],
+            "y": [4.4, 5.5, 6.6],
+            "t": [7.7, 8.8, 9.9]
+        }
+    }
+    """
+
+    def test_round_trip(self):
+        """
+        Test serialisation and deserialisation of a fully-defined
+        TableTrajectory.
+        """
+        instance = TableTrajectory(
+            attrs=TableTrajectory.Attrs(
+                x=[1.1, 2.2, 3.3], y=[4.4, 5.5, 6.6], t=[7.7, 8.8, 9.9]
+            )
+        )
+        assert_json_is_equal(
+            TestTableTrajectory.FULL_JSON, CODEC.dumps(instance)
+        )
+        assert instance == CODEC.loads(
+            TableTrajectory, TestTableTrajectory.FULL_JSON
         )